Relay: use 'relay.config.js' when fetching a new schema
The original idea of `relay.config.js` was that all the tools can use it and the configuration is only in one place. This commit does exactly that by allowing `adeira-fetch-schema` to read the values from this config. I still kept the previous CLI interface for backward compatibility though.

I also moved the validation logic closer to the CLI options resolver where it belongs.

adeira-source-id: a61a28afa39fdccb1a0be16f7c2b36fdbbecac17