Relay: update `RelayLogger`
This is essentially a continuation of 52115ad832860ef7dc259f6e43e232ca544ac65a.

It turns the "debug" logger into more useful logger by hiding some
unnecessary and spamy events and makes the other events more useful.
For example, it shows whether the data is available/stale/missing and
whether it's going to be fetched/re-fetched or not.

Other updates will most likely follow as we go.

adeira-source-id: f95f2aaa057124dad59fbdae07e59bcd839aa042